+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.7.0](https://github.com/americanexpress/fetchye/compare/v1.6.0...v1.7.0) (2025-07-15)
+
+
+### Bug Fixes
+
+* **readme:** update fetchClient required value ([#97](https://github.com/americanexpress/fetchye/issues/97)) ([ef4bcb5](https://github.com/americanexpress/fetchye/commit/ef4bcb573c0877ae7428e649671ecf9f127c6f93))
+
+
+### Features
+
+* force initial fetch ([#98](https://github.com/americanexpress/fetchye/issues/98)) ([e8e91d2](https://github.com/americanexpress/fetchye/commit/e8e91d26898e365a12480e43469a4ee8791cc915))
+
+
+
+
+
 # [1.6.0](https://github.com/americanexpress/fetchye/compare/v1.4.0...v1.6.0) (2024-08-29)